removed asset modal from manage assets button
